requirements.txt
absl-py~=0.10
astor==0.7.1
bleach==1.5.0
certifi==2018.8.24
chardet==3.0.4
Django==2.1.11
django-appconf==1.0.2
django-compressor==2.2
djangorestframework==3.9.1
flake8==3.5.0
gast==0.4.0
grpcio==1.37.0
html5lib==0.9999999
idna==2.7
Markdown==2.6.11
mccabe==0.6.1
numpy~=1.19.2
pika==0.12.0
Pillow==8.4.0
protobuf==3.9.2
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.5
rcssmin==1.0.6
requests==2.21.0
rjsmin==1.0.12
six==1.15.0
tensorboard~=2.6
tensorflow==2.6.0
tensorflow-hub==0.1.1
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.15.3
